# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API for managing an inventory of mobile devices.",
        "title": "Device Manager API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "url": "https://tiagos4ntos.github.io",
            "email": "tiago482@gmail.com"
        },
        "license": {
            "name": "APACHE 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "0.0.1-beta"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
        "/devices": {
            "get": {
                "description": "Get all devices",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "devices"
                ],
                "summary": "List devices",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/dto.DeviceResponse"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/errors.DefaultErrorResult"
                        }
                    }
                }
            },
            "post": {
                "description": "Registers a new device",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "devices"
                ],
                "summary": "Create a new device",
                "parameters": [
                    {
                        "description": "Device payload",
                        "name": "device",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.CreateDeviceRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/dto.DeviceResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/errors.DefaultErrorResult"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/errors.DefaultErrorResult"
                        }
                    }
                }
            }
        },
        "/devices/{id}": {
            "get": {
                "description": "Returns a single device by its ID",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "devices"
                ],
                "summary": "Get device by ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Device ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dto.DeviceResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/errors.DefaultErrorResult"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/errors.DefaultErrorResult"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/errors.DefaultErrorResult"
                        }
                    }
                }
            },
            "delete": {
                "description": "Removes a device by ID",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "devices"
                ],
                "summary": "Delete a device",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Device ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/errors.DefaultErrorResult"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/errors.DefaultErrorResult"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "dto.CreateDeviceRequest": {
            "type": "object",
            "required": [
                "brand",
                "name",
                "state"
            ],
            "properties": {
                "brand": {
                    "type": "string",
                    "example": "Motorola"
                },
                "name": {
                    "type": "string",
                    "example": "Moto G100"
                },
                "state": {
                    "type": "string",
                    "enum": [
                        "available",
                        "in-use",
                        "inactive"
                    ],
                    "example": "available"
                }
            }
        },
        "dto.DeviceResponse": {
            "type": "object",
            "properties": {
                "brand": {
                    "type": "string",
                    "example": "Apple"
                },
                "created_at": {
                    "type": "string",
                    "example": "2025-08-31T21:00:00Z"
                },
                "deleted_at": {
                    "type": "string",
                    "example": "null"
                },
                "id": {
                    "type": "string",
                    "example": "550e8400-e29b-41d4-a716-446655440000"
                },
                "name": {
                    "type": "string",
                    "example": "iPhone 13"
                },
                "state": {
                    "type": "string",
                    "example": "available"
                },
                "updated_at": {
                    "type": "string",
                    "example": "2025-08-31T21:00:00Z"
                }
            }
        },
        "errors.DefaultErrorResult": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string"
                }
            }
        }
    }
}